+++ packages/xchain-util/src/asset.test.ts
@@ -20,10 +20,22 @@ import {
   AssetBTC,
   AssetETH,
   AssetBNB,
+  isBigNumberValue,
 } from './asset'
 import { Asset, Denomination } from './types'
 
 describe('asset', () => {
+  describe('isBigNumberValue', () => {
+    it('return true for BigNumber.Value', () => {
+      expect(isBigNumberValue(1)).toBeTruthy()
+      expect(isBigNumberValue('1')).toBeTruthy()
+      expect(isBigNumberValue(bn(1))).toBeTruthy()
+    })
+    it('return false for others', () => {
+      expect(isBigNumberValue(baseAmount(10))).toBeFalsy()
+      expect(isBigNumberValue(assetAmount(10))).toBeFalsy()
+    })
+  })
   describe('assetAmount', () => {
     it('should create asset amount by given value', () => {
       const amount = assetAmount(10)
@@ -38,6 +50,68 @@ describe('asset', () => {
       expect(amount.amount()).toEqual(bn('10'))
       expect(amount.decimal).toEqual(18)
     })
+    it('should be able to add a value', () => {
+      const amount = assetAmount(10, 18).plus(100)
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('110'))
+    })
+    it('should be able to sub a value', () => {
+      const amount = assetAmount(10, 18).minus(5)
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('5'))
+    })
+    it('should be able to mul a value', () => {
+      const amount = assetAmount(10, 18).times(1e10)
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('100000000000'))
+    })
+    it('should be able to div a value', () => {
+      const amount = assetAmount(10, 18).div(5)
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('2'))
+    })
+    it('should be able to add AssetAmount', () => {
+      const amount = assetAmount(10, 18).plus(assetAmount(10, 10))
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('20'))
+      expect(amount.decimal).toEqual(18)
+    })
+    it('should be able to set decimal', () => {
+      const amount = assetAmount(10, 18).plus(assetAmount(10, 10), 12)
+      expect(amount.type).toEqual(Denomination.ASSET)
+      expect(amount.amount()).toEqual(bn('20'))
+      expect(amount.decimal).toEqual(12)
+    })
+    it('should be able to check lt', () => {
+      expect(assetAmount(10).lt(assetAmount(11))).toBeTruthy()
+      expect(assetAmount(10).lt(11)).toBeTruthy()
+      expect(assetAmount(11).lt(10)).toBeFalsy()
+      expect(assetAmount(11, 10).lt(assetAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check lte', () => {
+      expect(assetAmount(10).lte(10)).toBeTruthy()
+      expect(assetAmount(11).lte(10)).toBeFalsy()
+      expect(assetAmount(10).lte(assetAmount(10))).toBeTruthy()
+      expect(assetAmount(11, 10).lte(assetAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check gt', () => {
+      expect(assetAmount(10).gt(9)).toBeTruthy()
+      expect(assetAmount(9).gt(10)).toBeFalsy()
+      expect(assetAmount(10).gt(assetAmount(9))).toBeTruthy()
+      expect(assetAmount(9, 10).gt(assetAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check gte', () => {
+      expect(assetAmount(10).gte(10)).toBeTruthy()
+      expect(assetAmount(9).gte(10)).toBeFalsy()
+      expect(assetAmount(10).gte(assetAmount(10))).toBeTruthy()
+      expect(assetAmount(9, 10).gte(assetAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check eq', () => {
+      expect(assetAmount(10).eq(10)).toBeTruthy()
+      expect(assetAmount(9).eq(10)).toBeFalsy()
+      expect(assetAmount(10).eq(assetAmount(10))).toBeTruthy()
+      expect(assetAmount(9, 10).eq(assetAmount(10, 8))).toBeFalsy()
+    })
   })
 
   describe('baseAmount', () => {
@@ -45,6 +119,68 @@ describe('asset', () => {
       const amount = baseAmount(10)
       expect(amount.type).toEqual(Denomination.BASE)
       expect(amount.amount()).toEqual(bn('10'))
+    })
+    it('should be able to add a value', () => {
+      const amount = baseAmount(10).plus(100)
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('110'))
+    })
+    it('should be able to sub a value', () => {
+      const amount = baseAmount(10).minus(5)
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('5'))
+    })
+    it('should be able to mul a value', () => {
+      const amount = baseAmount(10).times(5)
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('50'))
+    })
+    it('should be able to div a value', () => {
+      const amount = baseAmount(10).div(5)
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('2'))
+    })
+    it('should be able to add BaseAmount', () => {
+      const amount = baseAmount(10, 18).plus(baseAmount(10, 10))
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('20'))
+      expect(amount.decimal).toEqual(18)
+    })
+    it('should be able to set decimal', () => {
+      const amount = baseAmount(10, 18).plus(baseAmount(10, 10), 12)
+      expect(amount.type).toEqual(Denomination.BASE)
+      expect(amount.amount()).toEqual(bn('20'))
+      expect(amount.decimal).toEqual(12)
+    })
+    it('should be able to check lt', () => {
+      expect(baseAmount(10).lt(11)).toBeTruthy()
+      expect(baseAmount(10).lt(9)).toBeFalsy()
+      expect(baseAmount(10).lt(baseAmount(11))).toBeTruthy()
+      expect(baseAmount(10, 10).lt(baseAmount(9, 8))).toBeFalsy()
+    })
+    it('should be able to check lte', () => {
+      expect(baseAmount(10).lte(10)).toBeTruthy()
+      expect(baseAmount(10).lte(9)).toBeFalsy()
+      expect(baseAmount(10).lte(baseAmount(10))).toBeTruthy()
+      expect(baseAmount(10, 10).lte(baseAmount(9, 8))).toBeFalsy()
+    })
+    it('should be able to check gt', () => {
+      expect(baseAmount(10).gt(9)).toBeTruthy()
+      expect(baseAmount(9).gt(10)).toBeFalsy()
+      expect(baseAmount(10).gt(baseAmount(9))).toBeTruthy()
+      expect(baseAmount(9, 10).gt(baseAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check gte', () => {
+      expect(baseAmount(10).gte(10)).toBeTruthy()
+      expect(baseAmount(9).gte(10)).toBeFalsy()
+      expect(baseAmount(10).gte(baseAmount(10))).toBeTruthy()
+      expect(baseAmount(9, 10).gte(baseAmount(10, 8))).toBeFalsy()
+    })
+    it('should be able to check eq', () => {
+      expect(baseAmount(10).eq(10)).toBeTruthy()
+      expect(baseAmount(9).eq(10)).toBeFalsy()
+      expect(baseAmount(10).eq(baseAmount(10))).toBeTruthy()
+      expect(baseAmount(9, 10).eq(baseAmount(10, 8))).toBeFalsy()
     })
   })