+++ packages/xchain-util/src/types/index.ts
@@ -15,6 +15,15 @@ export enum Denomination {
 type Amount<T> = {
   type: T
   amount: () => BigNumber
+  plus: (value: BigNumber.Value | Amount<T>, decimal?: number) => Amount<T>
+  minus: (value: BigNumber.Value | Amount<T>, decimal?: number) => Amount<T>
+  times: (value: BigNumber.Value | Amount<T>, decimal?: number) => Amount<T>
+  div: (value: BigNumber.Value | Amount<T>, decimal?: number) => Amount<T>
+  gt: (value: BigNumber.Value | Amount<T>) => boolean
+  gte: (value: BigNumber.Value | Amount<T>) => boolean
+  lt: (value: BigNumber.Value | Amount<T>) => boolean
+  lte: (value: BigNumber.Value | Amount<T>) => boolean
+  eq: (value: BigNumber.Value | Amount<T>) => boolean
   decimal: number
 }